+++ doc/changes.rst
@@ -5,7 +5,9 @@ desitarget Change Log
 1.3.1 (unreleased)
 ------------------
 
-* No changes yet.
+* Also use the ops/tiles-specstatus.ecsv tile file for SV [`PR #765`_].
+
+.. _`PR #765`: https://github.com/desihub/desitarget/pull/765
 
 1.3.0 (2021-09-20)
 ------------------